[Refactor]: Space 검색 API에서 동적 쿼리 생성 객체 분리 (DynamicQueryFactory)
