+++ src/main/java/com/tenten/linkhub/domain/space/controller/mapper/SpaceApiMapper.java
@@ -1,10 +1,9 @@
 package com.tenten.linkhub.domain.space.controller.mapper;
 
-import com.tenten.linkhub.domain.space.controller.dto.SpacesFindByQueryApiRequest;
+import com.tenten.linkhub.domain.space.controller.dto.space.SpacesFindByQueryApiRequest;
 import com.tenten.linkhub.domain.space.service.dto.SpacesFindByQueryRequest;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.springframework.data.domain.Pageable;
 
 @Mapper(componentModel = "spring",
