+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/DynamicQueryFactory.java
@@ -1,0 +1,53 @@
+package com.tenten.linkhub.domain.space.repository.space;
+
+import com.querydsl.core.types.NullExpression;
+import com.querydsl.core.types.Order;
+import com.querydsl.core.types.OrderSpecifier;
+import com.querydsl.core.types.dsl.BooleanExpression;
+import com.tenten.linkhub.domain.space.model.category.Category;
+import lombok.NoArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+
+import static com.tenten.linkhub.domain.space.model.space.QSpace.space;
+
+@Component // 1
+@NoArgsConstructor
+public class DynamicQueryFactory {
+
+    public OrderSpecifier<String> spaceSort(Pageable pageable) {
+        for (Sort.Order sort: pageable.getSort()){
+            String property = sort.getProperty();
+
+            switch (property){
+                case "favorite_count" -> {
+                    return new OrderSpecifier(Order.DESC, space.favoriteCount);
+                }
+                case "created_at"-> {
+                    return new OrderSpecifier(Order.DESC, space.createdAt);
+                }
+            }
+        }
+
+        return new OrderSpecifier(Order.ASC, NullExpression.DEFAULT, OrderSpecifier.NullHandling.Default);
+    }
+
+    public BooleanExpression eqCategory(Category filter) {
+        if (filter != null){
+            return space.category.eq(filter);
+        }
+
+        return null;
+    }
+
+    public BooleanExpression eqSpaceName(String keyWord) {
+        if (StringUtils.hasText(keyWord)){
+            return space.spaceName.contains(keyWord);
+        }
+
+        return null;
+    }
+
+}
