+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/SpaceQueryRepository.java
@@ -1,19 +1,11 @@
 package com.tenten.linkhub.domain.space.repository.space;
 
-import com.querydsl.core.types.NullExpression;
-import com.querydsl.core.types.Order;
-import com.querydsl.core.types.OrderSpecifier;
-import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.tenten.linkhub.domain.space.model.category.Category;
 import com.tenten.linkhub.domain.space.model.space.Space;
-import com.tenten.linkhub.domain.space.repository.space.dto.QueryCond;
-import org.springframework.data.domain.Pageable;
+import com.tenten.linkhub.domain.space.repository.space.dto.QueryCondition;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Repository;
-import org.springframework.util.StringUtils;
 
 import java.util.List;
 
@@ -23,20 +15,24 @@ import static com.tenten.linkhub.domain.space.model.space.QSpace.space;
 public class SpaceQueryRepository {
 
     private final JPAQueryFactory queryFactory;
+    private final DynamicQueryFactory dynamicQueryFactory;
 
-    public SpaceQueryRepository(JPAQueryFactory queryFactory) {
+    public SpaceQueryRepository(JPAQueryFactory queryFactory, DynamicQueryFactory dynamicQueryFactory) {
         this.queryFactory = queryFactory;
+        this.dynamicQueryFactory = dynamicQueryFactory;
     }
 
-    public Slice<Space> findByCondition(QueryCond condition){
+    public Slice<Space> findByCondition(QueryCondition condition){
+
         List<Space> spaces = queryFactory
                 .select(space)
                 .from(space)
                 .where(space.isDeleted.eq(false),
-                        eqSpaceName(condition.keyWord()),
-                        eqCategory(condition.filter())
+                        dynamicQueryFactory.eqSpaceName(condition.keyWord()),
+                        dynamicQueryFactory.eqCategory(condition.filter())
                 )
-                .orderBy(spaceSort(condition.pageable()))
+                .orderBy()
+                .orderBy(dynamicQueryFactory.spaceSort(condition.pageable()))
                 .offset(condition.pageable().getOffset())
                 .limit(condition.pageable().getPageSize() + 1)
                 .fetch();
@@ -49,39 +45,6 @@ public class SpaceQueryRepository {
         }
 
         return new SliceImpl<>(spaces, condition.pageable(), hasNext);
-    }
-
-    private OrderSpecifier<String> spaceSort(Pageable pageable) {
-        for (Sort.Order sort: pageable.getSort()){
-            String property = sort.getProperty();
-
-            switch (property){
-                case "favorite_count" -> {
-                    return new OrderSpecifier(Order.DESC, space.favoriteCount);
-                }
-                case "created_at"-> {
-                    return new OrderSpecifier(Order.DESC, space.createdAt);
-                }
-            }
-        }
-
-        return new OrderSpecifier(Order.ASC, NullExpression.DEFAULT, OrderSpecifier.NullHandling.Default);
-    }
-
-    private BooleanExpression eqCategory(Category filter) {
-        if (filter != null){
-            return space.category.eq(filter);
-        }
-
-        return null;
-    }
-
-    private BooleanExpression eqSpaceName(String keyWord) {
-        if (StringUtils.hasText(keyWord)){
-            return space.spaceName.contains(keyWord);
-        }
-
-        return null;
     }
 
 }
