+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/SpaceRepository.java
@@ -1,11 +1,9 @@
 package com.tenten.linkhub.domain.space.repository.space;
 
 import com.tenten.linkhub.domain.space.model.space.Space;
-import com.tenten.linkhub.domain.space.repository.space.dto.QueryCond;
+import com.tenten.linkhub.domain.space.repository.space.dto.QueryCondition;
 import org.springframework.data.domain.Slice;
 
-import java.util.List;
-
 public interface SpaceRepository {
-    Slice<Space> findByQuery(QueryCond queryCond);
+    Slice<Space> findByQuery(QueryCondition queryCondition);
 }
