+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/SpaceRepositoryImpl.java
@@ -1,11 +1,9 @@
 package com.tenten.linkhub.domain.space.repository.space;
 
 import com.tenten.linkhub.domain.space.model.space.Space;
-import com.tenten.linkhub.domain.space.repository.space.dto.QueryCond;
+import com.tenten.linkhub.domain.space.repository.space.dto.QueryCondition;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public class SpaceRepositoryImpl implements SpaceRepository{
@@ -19,8 +17,8 @@ public class SpaceRepositoryImpl implements SpaceRepository{
     }
 
     @Override
-    public Slice<Space> findByQuery(QueryCond queryCond) {
-        return spaceQueryRepository.findByCondition(queryCond);
+    public Slice<Space> findByQuery(QueryCondition queryCondition) {
+        return spaceQueryRepository.findByCondition(queryCondition);
     }
 
 }
