+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/dto/QueryCondition.java
@@ -3,7 +3,7 @@ package com.tenten.linkhub.domain.space.repository.space.dto;
 import com.tenten.linkhub.domain.space.model.category.Category;
 import org.springframework.data.domain.Pageable;
 
-public record QueryCond(
+public record QueryCondition(
         Pageable pageable,
         String keyWord,
         Category filter
